Fixed Bugs at EspeciesRepository Query

+++ src/main/java/com/oceanguardians/oceanguard/services/EspeciesService.java
@@ -6,6 +6,6 @@ import com.oceanguardians.oceanguard.model.Veda;
 import java.util.List;
 
 public interface EspeciesService {
-    List<Especies> buscarEspeciesPorRegion(long idRegion);
+    List<Especies> buscarEspeciesPorRegion(int idRegion);
 
 }
